[hotfix] Introduce TaskManagerExceptionUtils
