+++ flink-core/src/test/java/org/apache/flink/util/ExceptionUtilsTest.java
@@ -91,7 +91,7 @@ public class ExceptionUtilsTest extends TestLogger {
 
 	@Test
 	public void testTryEnrichTaskExecutorErrorCanHandleNullValue() {
-		assertThat(ExceptionUtils.tryEnrichTaskManagerError(null), is(nullValue()));
+		assertThat(ExceptionUtils.tryEnrichOutOfMemoryError(null, "", ""), is(nullValue()));
 	}
 
 	@Test
